Move interfaces to utils
