+++ src/magic_tiler/utils/interfaces.py
